Migrate rspec test to Test::Unit
